add to cart working
so we are nearly at the end of the mern project. We can add the product to the cart and also there is Badge in the CustomButtons file, when the user add a product to cart, we can see how many products are added to the cart. It is called badge.